Changed 'process_1REP_DS_data' to use 'ts_avg_reps'
Figured out that I'm basically doing the same thing as Lisa in 'process_data'. The main thing to be aware of now is that I am using the generic 'timestamps' for each condition. This is what the timing should be. Whereas, before I was getting the timing directly from Log.ADC.Time and parsing it based on 'cond_start_times'.